packages/perfectprint-ai/processor/requirements.txt
# FastAPI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing
pillow>=10.4.0
numpy>=1.26.2
opencv-python-headless>=4.8.1.78

# Background Removal - BRIA-RMBG-2.0
transformers>=4.35.2
torch>=2.1.1
torchvision>=0.16.1

# Vectorization - VTracer
vtracer>=0.6.10

# Upscaling - Real-ESRGAN (AI upscaling for fixing pixelation)
realesrgan==0.3.0
basicsr>=1.4.2

# Utilities
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

